test(cli): split sync package coverage
